requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==5.0.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-allauth==0.61.1
django-autocomplete-light==3.9.7
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-heroku==0.3.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-paypal==2.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.50.0
gunicorn==22.0.0
html5lib==1.1
idna==3.6
instamojo-wrapper==1.2.0
kiwisolver==1.4.5
kombu==5.3.5
lxml==5.1.0
matplotlib==3.8.4
mysql-connector-python==8.1.0
mysqlclient==2.2.1
numpy==1.26.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.0
pandas==2.2.1
pdfdocument==4.0.0
pdfkit==1.0.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psycopg2==2.9.9
pycparser==2.21
pydyf==0.9.0
pyHanko==0.23.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.1.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.4.0
seaborn==0.13.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==8.2.0
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
zopfli==0.2.3
